Extract home components files
