+++ src/components/home/Photo.tsx
