+++ src/components/home/RecentFightBox.tsx
